% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/efaR.R
\name{efaR}
\alias{efaR}
\title{Wrapper function for EFA XML API.}
\usage{
efaR(origin, destination, time = Sys.time(), simplify = TRUE)
}
\arguments{
\item{origin}{The origin of the journey. Either a name or a DHID Code. If a name is provided the next station is calculated based on the geocode function and the 'Haversine distance'.}

\item{destination}{The destination of the journey. Either a name or a DHID Code. If a name is provided the next station is calculated based on the geocode function and the 'Haversine distance'.}

\item{time}{The date and time of the journey.}

\item{simplify}{Simplify determines whether the original xml result is returned or a simpler S3 object.}
}
\value{
Either the original xml result or an S3 object based on the simplify input.
}
\description{
Wrapper function for EFA XML API.
}
